Merged PR 20704: User Story 34536: MSTP configuration - config page

User Story 34536: MSTP configuration - config page

Related work items: #34536

Related work items: #34536

+++ app/src/main/java/a75f/io/renatus/SystemFragment.java
@@ -1727,8 +1727,8 @@ public class SystemFragment extends Fragment implements AdapterView.OnItemSelect
 
 				if(presentationData != null && valueConstraint != null){
 					if(valueConstraint.getMinValue() != null && valueConstraint.getMinValue() != null){
-						int minValue = valueConstraint.getMinValue();
-						int maxValue = valueConstraint.getMaxValue();
+						int minValue = valueConstraint.getMinValue().intValue();
+						int maxValue = valueConstraint.getMaxValue().intValue();
 						//float step = Float.parseFloat(presentationData.getTagValueIncrement(
 						double step = Double.parseDouble(presentationData.getTagValueIncrement());
 						spinnerValues = generateValuesForSpinner(minValue, maxValue, step);
